Moving all the utilities used for testing existing apps into a more logical location
